query node mappings restructuralization
