+++ query-node/mappings/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  env: {
+    node: true,
+  },
+  rules: {},
+}
