+++ query-node/mappings/src/common.ts
