+++ query-node/mappings/src/genesis-data/index.ts
