+++ query-node/mappings/src/genesis-data/membershipSystem.json
