+++ query-node/mappings/src/genesis-data/workers.json
