+++ query-node/mappings/src/genesis-data/workingGroups.json
