+++ query-node/mappings/src/genesis.ts
