+++ query-node/mappings/src/index.ts
