+++ query-node/mappings/tsconfig.json
@@ -21,5 +21,5 @@
       // "query-node/*": [ "../generated/graphql-server/src/*" ]
     }
   },
-  "include": ["./**/*"]
+  "include": ["./src/**/*"]
 }
